fix(ci): python-full-coverage — decompose four files over the 1500-line maintainability limit
test_first_party_code_files_stay_under_line_limit flagged four first-party
files exceeding MAX_FIRST_PARTY_FILE_LINES (1500). Split each below the limit
by relocating cohesive, behavior-preserving blocks into sibling modules/files,
following the repo's existing decomposition conventions:

- src/awf/api/schemas.py (1562->1371): leaf request/response models moved to
  new schemas_responses.py and re-exported so the import surface is unchanged.
- src/awf/service/gc.py (1669->1491): leaf path/classification helpers moved to
  new gc_classify.py and re-imported so public/private names stay on awf.service.gc.
- tests/unit/node/test_compose_manager.py (1710->1335): TestCompanionImageCommands
  and the docker-capture tests moved verbatim to test_compose_manager_companions.py.
- tests/unit/service/test_gc_parts/test_gc_part_001.py (1526->1332): trailing
  tests moved verbatim to a new test_gc_part_003.py.

No test, assertion, or the 1500 constant was weakened. Verified: ruff, ruff
format, mypy (full src), OpenAPI drift gate, the maintainability suite, and the
affected gc/compose/api targeted suites (1134 passed).
